+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Automatically patch a version for each commit
 - Allow to specify manual version in publish script
 - Automatically deploy meetingV2 to serverless demo
+- Realtime controller now exposes external user ID in places where attendee ID is present
 
 ### Fixed
 - Remove line endings in the keyword when searching for connection attributes in SDP